+++ ChangeLog.md
@@ -1,3 +1,7 @@
+- v1.0.4
+  - 使用 [RabbitCommon v2.2.0](https://github.com/KangLin/RabbitCommon/releases/tag/v2.2.0)
+  - 删除 vcpkg.json 中 baseline
+
 - v1.0.3
   - 使用 [RabbitCommon v2.1.0](https://github.com/KangLin/RabbitCommon/releases/tag/v2.1.0)